requirements.txt
streamlit>=1.28.0
pypdf2>=3.0.0
python-docx>=1.1.0
pandas>=2.0.0
openai>=1.3.0
python-dotenv>=1.0.0 